replace deliveredDataDeleter with new one  also point requirement to EGCG-Core v0.5
