+++ requirements.txt
@@ -1,3 +1,3 @@
-git+https://github.com/EdinburghGenomics/EGCG-Core.git@v0.4.5
+git+https://github.com/EdinburghGenomics/EGCG-Core.git@v0.5
 jinja2==2.8
 weasyprint==0.30
